pyproject.toml
[project]
name = "agent-without-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.78.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "wikipedia>=1.4.0",
]
